# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, kb, memory, data, file, plugin, mcp, skill. The
    config scope also grants bot and provider access.
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/files:
    post:
      tags: [Files]
      summary: Upload a file
      operationId: uploadFile
      x-astrbot-scope: file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/file:
    post:
      tags: [Open API]
      summary: Upload a file
      operationId: uploadOpenApiFile
      x-astrbot-scope: file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    get:
      tags: [Open API]
      summary: Download an uploaded file by attachment ID
      operationId: downloadOpenApiFile
      x-astrbot-scope: file
      parameters:
        - $ref: '#/components/parameters/AttachmentIdQuery'
      responses:
        '200':
          description: File content or an error envelope
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
  /api/v1/files/{attachment_id}:
    get:
      tags: [Files]
      summary: Get attachment metadata
      operationId: getAttachment
      x-astrbot-scope: file
      parameters:
        - $ref: '#/components/parameters/AttachmentId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
    delete:
      tags: [Files]
      summary: Delete an attachment
      operationId: deleteAttachment
      x-astrbot-scope: file
      parameters:
        - $ref: '#/components/parameters/AttachmentId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
  /api/v1/plugins/{extension_id}/dashboard/files/{action_id}:
    post:
      tags: [Plugin Dashboard]
      summary: Invoke a Dashboard extension file Action and create a ticket
      operationId: invokePluginDashboardFile
      security:
        - DashboardBearerAuth: []
          DashboardCookieAuth: []
      parameters:
        - $ref: '#/components/parameters/ExtensionId'
        - $ref: '#/components/parameters/PluginActionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginDashboardFileRequest'
      responses:
        '200':
          description: File ticket created
          headers:
            Set-Cookie:
              description: Exact-Path HttpOnly file-ticket secret cookie.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginDashboardFileResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
        '415':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '504':
          $ref: '#/components/responses/Error'
...
  /api/v1/skills/{skill_name}/files:
    get:
      tags: [Skills]
      summary: List files in a skill
      operationId: listSkillFiles
      x-astrbot-scope: skill
      parameters:
        - $ref: '#/components/parameters/SkillName'
        - name: path
          in: query
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/Ok'
  /api/v1/skills/{skill_name}/files/{file_path}:
    get:
      tags: [Skills]
      summary: Get skill file content
      operationId: getSkillFile
      x-astrbot-scope: skill
      parameters:
        - $ref: '#/components/parameters/SkillName'
        - $ref: '#/components/parameters/FilePath'
      responses:
        '200':
          $ref: '#/components/responses/Text'
    put:
      tags: [Skills]
      summary: Update skill file content
      operationId: updateSkillFile
      x-astrbot-scope: skill
      parameters:
        - $ref: '#/components/parameters/SkillName'
        - $ref: '#/components/parameters/FilePath'
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
  /api/v1/knowledge-bases/{kb_id}/stats:
    get:
      tags: [Knowledge Base]
      summary: Get knowledge base stats
      operationId: getKnowledgeBaseStats
      x-astrbot-scope: kb
      parameters:
        - $ref: '#/components/parameters/KbId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
  /api/v1/stats:
    get:
      tags: [Stats]
      summary: Get runtime statistics
      operationId: getStats
      x-astrbot-scope: system
      parameters:
        - name: offset_sec
          in: query
          schema:
            type: integer
            default: 86400
      responses:
        '200':
          $ref: '#/components/responses/Ok'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DashboardBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Dashboard session JWT
    DashboardCookieAuth:
      type: apiKey
      in: cookie
      name: astrbot_dashboard_jwt
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
    Error:
      description: Standard AstrBot error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Text:
      description: Text response
      content:
        text/plain:
          schema:
            type: string
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
...
  parameters:
    ExtensionId:
      name: extension_id
      in: path
      required: true
      schema:
        type: string
        minLength: 3
        maxLength: 128
        pattern: '^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?(?:\.[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?)+$'
...
    PluginActionId:
      name: action_id
      in: path
      required: true
      schema:
        type: string
        maxLength: 64
        pattern: '^[a-z][a-z0-9]*(?:[._-][a-z0-9]+)*$'
    AttachmentId:
      name: attachment_id
      in: path
      required: true
      schema:
        type: string
    AttachmentIdQuery:
      name: attachment_id
      in: query
      required: true
      schema:
        type: string
...
    FilePath:
      name: file_path
      in: path
      required: true
      description: URL-encoded relative file path.
      schema:
        type: string
...
    KbId:
      name: kb_id
      in: path
      required: true
      schema:
        type: string
...
    SkillName:
      name: skill_name
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
    ErrorEnvelope:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          const: error
        message:
          type: string
        data: {}
      additionalProperties: false
...
    PluginDashboardFileRequest:
      type: object
      required:
        [
          protocol_version,
          instance_id,
          expected_generation,
          expected_disposition,
          payload,
        ]
      properties:
        protocol_version:
          type: integer
          const: 1
        instance_id:
          type: string
          minLength: 1
          maxLength: 128
        expected_generation:
          type: string
          minLength: 1
          maxLength: 128
        expected_disposition:
          type: string
          enum: [inline, attachment]
        payload: {}
      additionalProperties: false
...
    PluginDashboardFileTicket:
      type: object
      required:
        [ticket_url, filename, content_type, size, disposition, expires_at]
      properties:
        ticket_url:
          type: string
        filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
          minimum: 0
        disposition:
          type: string
          enum: [inline, attachment]
        expires_at:
          type: string
          format: date-time
      additionalProperties: false
    PluginDashboardFileResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: [string, 'null']
        data:
          $ref: '#/components/schemas/PluginDashboardFileTicket'
      additionalProperties: false
...
    FileUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
...